Consume store procedure for adding new photo, upload photo to blob

+++ Pada/Controllers/PendingPhotoRequestsController.cs
@@ -1,14 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.Data.SqlClient;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Blob;
@@ -44,6 +47,10 @@ namespace Pada.Views
             {
                 return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
             }
+            if(user.UserLevel <4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             ViewBag.CurrentUserEmail = user.Email;
             return View(await _context.PendingPhotoRequest.ToListAsync());
         }
@@ -51,11 +58,15 @@ namespace Pada.Views
         // GET: PendingPhotoRequests/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var user = await _userManager.GetUserAsync(User);
             if (id == null)
             {
                 return NotFound();
             }
-
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             var pendingPhotoRequest = await _context.PendingPhotoRequest
                 .FirstOrDefaultAsync(m => m.TransactionId == id);
             if (pendingPhotoRequest == null)
@@ -77,49 +88,63 @@ namespace Pada.Views
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("TransactionId,Email,Type,OldPhotoPath,NewPhotoPath,AcceptedVote,RejectedVote,VoteRequired,FullPhotoPath,FacePhotoPath,AnyPhotoPath,PendingVote,ImageName,ImageFile")] PendingPhotoRequest pendingPhotoRequest)
+        public async Task<IActionResult> Create([Bind("TransactionId,Email,Type,OldPhotoPath,NewPhotoPath,AcceptedVote,RejectedVote,VoteRequired,FullPhotoPath,FacePhotoPath,AnyPhotoPath,PendingVote,ImageName,ImageFile")] PendingPhotoRequest pendingPhotoRequest, IFormFile fullImageFile, IFormFile faceImageFile, IFormFile anyImageFile)
         {
+            var user = await _userManager.GetUserAsync(User);
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             if (ModelState.IsValid)
             {
                 //Save image to wwwroot/image
-                var user = await _userManager.GetUserAsync(User);
                 if (user == null)
                 {
                     return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
                 }
-                string wwwRootPath = _hostEnvironment.WebRootPath;
-                string fileName = Path.GetFileNameWithoutExtension(pendingPhotoRequest.ImageFile.FileName);
-                string extension = Path.GetExtension(pendingPhotoRequest.ImageFile.FileName);
-                pendingPhotoRequest.ImageName = fileName = fileName + DateTime.Now.ToString("yymmssfff") + extension;
-                //Get current user email
-                pendingPhotoRequest.Email = user.Email;
-                string path = Path.Combine(wwwRootPath + "/image/", fileName);
-                using (var filestream = new FileStream(path, FileMode.Create))
-                {
-                    await pendingPhotoRequest.ImageFile.CopyToAsync(filestream);
-                }
-                string storageConnectionString = "DefaultEndpointsProtocol=https;AccountName=chinhlanhuvaycf7e180a308;AccountKey=ufsUzB/+lb4VQX2QDXECQO9ZRVWaxc2G6nKEQD1Bo40Px7n9Ajs5pJHWB0nmKfVFZnvaFaacsh/Trmtrci+U+w==;EndpointSuffix=core.windows.net";
-                CloudStorageAccount storageacc = CloudStorageAccount.Parse(storageConnectionString);
-
-                CloudBlobClient blobClient = storageacc.CreateCloudBlobClient();
-                CloudBlobContainer container = blobClient.GetContainerReference("vhds");
-                await container.CreateIfNotExistsAsync();
-
-                CloudBlockBlob blockBlob = container.GetBlockBlobReference(pendingPhotoRequest.ImageName);
-
-                using (var filestream = System.IO.File.OpenRead(path))
-                {
-                    await blockBlob.UploadFromStreamAsync(filestream);
-                }
-                //using (var memoryStream = new MemoryStream())
+                //string wwwRootPath = _hostEnvironment.WebRootPath;
+                //string fileName = Path.GetFileNameWithoutExtension(fullImageFile.FileName);
+                //string extension = Path.GetExtension(pendingPhotoRequest.ImageFile.FileName);
+                //pendingPhotoRequest.ImageName = fileName = fileName + DateTime.Now.ToString("yymmssfff") + extension;
+                ////Get current user email
+                //pendingPhotoRequest.Email = user.Email;
+                //string path = Path.Combine(wwwRootPath + "/image/", fileName);
+                //using (var filestream = new FileStream(path, FileMode.Create))
                 //{
-                //    await pendingPhotoRequest.ImageFile.CopyToAsync(memoryStream);
-                //    await blockBlob.UploadFromStreamAsync(memoryStream);
+                //    await pendingPhotoRequest.ImageFile.CopyToAsync(filestream);
                 //}
-                var blobUrl = blockBlob.Uri.AbsoluteUri;
-                pendingPhotoRequest.ImageName = blobUrl;
-                _context.Add(pendingPhotoRequest);
-                await _context.SaveChangesAsync();
+                //string storageConnectionString = "DefaultEndpointsProtocol=https;AccountName=chinhlanhuvaycf7e180a308;AccountKey=ufsUzB/+lb4VQX2QDXECQO9ZRVWaxc2G6nKEQD1Bo40Px7n9Ajs5pJHWB0nmKfVFZnvaFaacsh/Trmtrci+U+w==;EndpointSuffix=core.windows.net";
+                //CloudStorageAccount storageacc = CloudStorageAccount.Parse(storageConnectionString);
+
+                //CloudBlobClient blobClient = storageacc.CreateCloudBlobClient();
+                //CloudBlobContainer container = blobClient.GetContainerReference("vhds");
+                //await container.CreateIfNotExistsAsync();
+
+                //CloudBlockBlob blockBlob = container.GetBlockBlobReference(pendingPhotoRequest.ImageName);
+
+                //using (var filestream = System.IO.File.OpenRead(path))
+                //{
+                //    await blockBlob.UploadFromStreamAsync(filestream);
+                //}
+                //var blobUrl = blockBlob.Uri.AbsoluteUri;
+                //pendingPhotoRequest.ImageName = blobUrl;
+                //_context.Add(pendingPhotoRequest);
+                //await _context.SaveChangesAsync();
+                string fullImageLink = await pendingPhotoRequest.UploadPhoto(fullImageFile, _hostEnvironment);
+                string faceImageLink = await pendingPhotoRequest.UploadPhoto(faceImageFile, _hostEnvironment);
+                string anyImageLink = await pendingPhotoRequest.UploadPhoto(anyImageFile, _hostEnvironment);
+                var updateContext = new Pada_DataContext();
+                //var email = new SqlParameter("@email", user.Email);
+                //var password = new SqlParameter("@password", null);
+                //var fullphoto = new SqlParameter("@fullphoto", fullImageLink);
+                //var facephoto = new SqlParameter("@facephoto", faceImageLink);
+                //var anyphoto = new SqlParameter("@anyphoto", anyImageLink);
+                var output = new SqlParameter();
+                output.ParameterName = "@result";
+                output.SqlDbType = SqlDbType.Int;
+                output.Direction = ParameterDirection.Output;
+                updateContext.Database.ExecuteSqlInterpolated($"EXEC [dbo].[usp_minh_photorequest_createnewphoto] @email={user.Email},@password={null},@fullphoto={fullImageLink},@facephoto ={faceImageLink},@anyphoto={anyImageLink}, @result = {output} OUT");
+                
                 return RedirectToAction(nameof(Index));
             }
             return View(pendingPhotoRequest);
@@ -129,6 +154,11 @@ namespace Pada.Views
         // GET: PendingPhotoRequests/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            var user = await _userManager.GetUserAsync(User);
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -149,6 +179,11 @@ namespace Pada.Views
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("TransactionId,Email,Type,OldPhotoPath,NewPhotoPath,AcceptedVote,RejectedVote,VoteRequired,FullPhotoPath,FacePhotoPath,AnyPhotoPath,PendingVote")] PendingPhotoRequest pendingPhotoRequest)
         {
+            var user = await _userManager.GetUserAsync(User);
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             if (id != pendingPhotoRequest.TransactionId)
             {
                 return NotFound();
@@ -180,6 +215,11 @@ namespace Pada.Views
         // GET: PendingPhotoRequests/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            var user = await _userManager.GetUserAsync(User);
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -200,6 +240,11 @@ namespace Pada.Views
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var user = await _userManager.GetUserAsync(User);
+            if (user.UserLevel < 4)
+            {
+                return NotFound($"User with this ID '{_userManager.GetUserId(User)}' can not see the content of this page.");
+            }
             var pendingPhotoRequest = await _context.PendingPhotoRequest.FindAsync(id);
             _context.PendingPhotoRequest.Remove(pendingPhotoRequest);
             await _context.SaveChangesAsync();
